Changes for Tabs
Generic Tabs will have getters in the Advanced Java section of the agent for both the Group and the object (sliders) inside the group. This is needed because we have spread them over multiple ALP files. The Group is needed to set the visibility of the tab you want. The other getters are so that you can update them from the Interface.