Main slider reworks ate (#18)
* Slider reworks: All sliders are now fully compatible with the companyUI and vice versa.

Slider reworks: All sliders are now fully compatible with the companyUI and vice versa. Sliders for electricity (heatpumps, solar panels and ev, can not be set below their starting point, savings sliders (for mobility, electricity and heat) now influence the companyUI as well (but not the other way around (intentional!)). Overall much better quality of life for the user than it was before! Also adjustments like.
The sliders should not have been influenced in use for projects without companyUI (notice i say companyUI, cause for businessparks it is also possible to disable them). Furthermore, paused GC are now better integrated in the model and no longer cause errors in sliders as well.
When there is a GC that is paused: its companyUI can not be found by the user now and even if it is possible to find them still: it is disabled from use, should be unpaused first.

* Optional use of calculated profile maximum