Descriptions for Sliders in Tabs
Due to inconsistencies in AnyLogic's presentations of single agents vs populations I have decided its not worth continuing to try and determine all positions automatically.
Hence, the function f_setInfoText() now needs the position of the info bubble as an argument.
Also added descriptions for sliders in the electricity, heating and mobility tabs.