Company / Houses PV Sliders
For companies the PV Slider scales with the total potential (based on roof surface area).
For houses the slider is still just a percentage of households that have PV on their roof. If they do the amount is still randomised, which means this slider is still inconsistent, but for a large amount of houses this will average out. It is not a reasonable to assume that the amount of PV for households scales linearly with roof surface area, most houses determine their amount of PV based on the expected self consumption, which is usually the restricting factor