Setup energy hub dashboard + slider adjustments (#144)
* E-Hub_Dashboard

New Agent for dashboard, has a presentation on the main interface and the GIS map will be moved on top of this presentation

* E-Hub_Dashboard

New Agent for e-hub dashboard, has a presentation on the main interface and the GIS map is moved on top of this when navigating to it.
- Initial configuration button and setup is working: for now a button has been made in the e-hub tab, that can be selected, which puts the model into 'e-hub configuration mode'. Here you can use the filters to make a selection of gc, that will be used to use as the selection for the energyHub dashboard after clicking complete.
- next step: adding resultsUI and tabs to the e-hub dashboard.

* next developments for the energyHub dashboard: ResultsUI and ui tabs now connected properly

-> Now, only making them function correctly still needs to happen.
Also additional layout work has been performed. Looks much better now. But still in progress.

* updates to energyhub loadin

* bugfixes and updates

* Merge branch 'main' into Setup-EnergyHub_Dashboard + bugfixes after merge

* Updates for the ui-tabs maintainability.

Start on making the setting of the sliders based on the engine state more modular, so it can be used in multiple dashboards, and overrided projects.

* bugfixes after merge

* cleanup of agents and classes in maps

* fixes after merge

* slider updates and cleanup of update functionalities

* slider updates

* slider work

* final fixes slider update

* Removed c_selectedEnergyHubGC, added getter for button createEnergyHub.

---------

Co-authored-by: Luc-Sol <luc@solidegroep.nl>
Co-authored-by: GillisHommen <98752265+GillisHommen@users.noreply.github.com>