avgc_data update -> Gas heating value is now 9.77 (from 9.7) for more accuracy. Gasburner efficiency is now 0.94 (from 0.99)
Gas heating value was missing a decimal value, which has been added now, so it is now according most sources about the dutch gas heating values.
Also: Gasburner efficiency has been lowerd to 0.94, it was 0.99, this had no source yet. After research we found sources that gave 100-104 % but we use the higher hbut the energy density used was the lower heating value, we use the higher heating value in our model, so it should be according to that -> it becomes 0.94.