slider gc and slider bugfixes: slider gc population of main interface and ehub dashboard are managed correctly now.
Also: when activating a slider in ehub, it stays active when going to the main interface, but can not be adjusted from the main interface. It is representedin the minimum value of the main interface slider: works for pv farm, wind farm, and grid battery. Also: PV filter no longer selects PV gc that have no gis object (invisible slider gc)